int -> Long 타입으로 변경, 기업목록 작업 진행중

+++ src/main/java/com/jobgoalin/workinfo/info/AccessInfo.java
@@ -16,7 +16,7 @@ import java.sql.Timestamp;
 public class AccessInfo {
 
     @Id
-    private int accessLevel;
+    private Long accessLevel;
 
     private String accessDesc;
     private String instId;
